Refactor Stock Model
